+++ app/src/main/java/com/cornellappdev/uplift/ui/components/goalsetting/DeleteDialog.kt
@@ -1,0 +1,135 @@
+package com.cornellappdev.uplift.ui.components.goalsetting
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.window.Dialog
+import com.cornellappdev.uplift.R
+import com.cornellappdev.uplift.util.PRIMARY_BLACK
+import com.cornellappdev.uplift.util.montserratFamily
+
+/**
+ * @param onConfirm callback for when the user confirms the deletion
+ * @param onDismiss callback for when the user dismisses the dialog
+ * @return DeleteDialog composable that prompts the user to confirm the deletion of a reminder
+ */
+@Composable
+fun DeleteDialog(
+    onConfirm: () -> Unit,
+    onDismiss: () -> Unit
+) {
+    Dialog(
+        onDismissRequest = onDismiss
+    ) {
+        Card(
+            colors = CardDefaults.cardColors(
+                containerColor = Color.White
+            ),
+            modifier = Modifier
+                .width(250.dp)
+                .height(268.dp),
+            shape = RoundedCornerShape(16.dp),
+        ) {
+            Box(
+                modifier = Modifier.fillMaxSize(),
+            ) {
+                Icon(
+                    imageVector = Icons.Default.Close,
+                    contentDescription = "Close",
+                    tint = PRIMARY_BLACK,
+                    modifier = Modifier
+                        .size(32.dp)
+                        .offset(x = 206.dp, y = 12.dp)
+                        .clickable(
+                            interactionSource = remember { MutableInteractionSource() },
+                            indication = null
+                        ) {
+                            onDismiss()
+                        }
+                )
+                Column(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(horizontal = 16.dp, vertical = 24.dp),
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                    verticalArrangement = Arrangement.spacedBy(20.dp)
+                ) {
+
+                    Icon(
+                        painter = painterResource(id = R.drawable.ic_trash),
+                        contentDescription = "Delete",
+                        modifier = Modifier.size(52.dp)
+                    )
+                    Text(
+                        text = "Are you sure you want to delete this reminder?",
+                        fontFamily = montserratFamily,
+                        textAlign = TextAlign.Center,
+                        fontWeight = FontWeight.Bold
+                    )
+                    Surface(
+                        shape = RoundedCornerShape(38.dp),
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .height(42.dp)
+                            .padding(horizontal = 8.dp),
+                        color = Color.Black,
+                        shadowElevation = 6.dp,
+                        onClick = onConfirm,
+                    ) {
+                        Text(
+                            text = "Yes, I'm sure",
+                            fontFamily = montserratFamily,
+                            fontSize = 14.sp,
+                            color = Color.White,
+                            fontWeight = FontWeight.Bold,
+                            modifier = Modifier.padding(top = 12.dp),
+                            textAlign = TextAlign.Center
+                        )
+                    }
+
+                    Text(
+                        text = "Cancel",
+                        fontFamily = montserratFamily,
+                        fontSize = 14.sp,
+                        color = PRIMARY_BLACK,
+                        fontWeight = FontWeight.Bold,
+                        textAlign = TextAlign.Center,
+                        modifier = Modifier.clickable(onClick = onDismiss)
+                    )
+
+                }
+            }
+
+
+        }
+
+    }
+
+}
