+++ app/src/main/java/com/cornellappdev/uplift/ui/components/goalsetting/EditReminderSection.kt
@@ -1,61 +1,72 @@
 package com.cornellappdev.uplift.ui.components.goalsetting
 
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.animateDpAsState
-import androidx.compose.animation.expandVertically
-import androidx.compose.animation.shrinkVertically
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.CheckboxDefaults
-import androidx.compose.material3.ElevatedCard
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.min
 import androidx.compose.ui.unit.sp
-import com.cornellappdev.uplift.ui.components.general.DayOfWeekBar
-import com.cornellappdev.uplift.ui.components.general.ReminderSwitch
+import com.cornellappdev.uplift.R
 import com.cornellappdev.uplift.ui.screens.Reminder
-import com.cornellappdev.uplift.util.GRAY04
-import com.cornellappdev.uplift.util.GRAY08
-import com.cornellappdev.uplift.util.PRIMARY_YELLOW
+import com.cornellappdev.uplift.util.PRIMARY_BLACK
 import com.cornellappdev.uplift.util.montserratFamily
 
+/**
+ * EditReminderSection is a composable that allows the user to edit a reminder.
+ * @param reminder the reminder to edit
+ * @param addNewReminderState whether the reminder is new
+ * @param onReminderSave callback for when the reminder is saved
+ * @param openDelete callback for when the delete button is clicked
+ * @sample EditReminderSection(
+ *    reminder = Reminder(time = "9:00 AM", days = setOf("W"), enabled = true),
+ *    addNewReminderState = false,
+ *    onReminderSave = { reminder -> println(reminder) },
+ *    openDelete = { println("Delete") }
+ *    )
+ */
 @Composable
 fun EditReminderSection(
+    reminder: Reminder? = null,
     addNewReminderState: Boolean = false,
-    onAddNewReminderStateChange: (Boolean) -> Unit = {},
-    reminders: List<Reminder> = emptyList(),
-    onRemindersChange: (List<Reminder>) -> Unit = {}
+    onReminderSave: (Reminder) -> Unit = {},
+    openDelete: () -> Unit = {}
 ) {
-    var selectedDays by remember { mutableStateOf(setOf("W")) }
-    var checkedState by remember { mutableStateOf(false) }
-    var setTimeOn by remember { mutableStateOf(false) }
-    var hour by remember { mutableStateOf(9) }
-    var minute by remember { mutableStateOf("00") }
-    var isAm by remember { mutableStateOf(true) }
+    val reminderTimeParams = reminder?.time?.split(":")
+    val reminderTimeMinMeridian = reminderTimeParams?.get(1)?.split(" ")
+    var selectedDays by remember { mutableStateOf(reminder?.days ?: setOf("W")) }
+    var checkedState by remember {
+        mutableStateOf(reminder != null && reminder.days.size == 7)
+    }
+    var setTimeOn by remember { mutableStateOf(reminder != null) }
+    var hour by remember { mutableIntStateOf(reminderTimeParams?.get(0)?.toInt() ?: 9) }
+    var minute by remember { mutableStateOf(reminderTimeMinMeridian?.get(0) ?: "00") }
+    var isAm by remember {
+        mutableStateOf(if (reminder != null) reminderTimeMinMeridian?.get(1) == "AM" else true)
+    }
     Column(
         verticalArrangement = Arrangement.spacedBy(8.dp),
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
         EditReminderCard(
+            reminder = reminder,
             selectedDays = selectedDays,
             onDaySelected = { selectedDays = it },
             checkedState = checkedState,
@@ -78,216 +89,99 @@ fun EditReminderSection(
                 8.dp, Alignment.CenterHorizontally
             )
         ) {
-            DeleteButton(onClick = {})
+            DeleteButton(onClick = openDelete)
             SubmitReminderButton(onClick = {
-                val newReminder = Reminder(time = "$hour:$minute ${if (isAm) "AM" else "PM"}",
-                    days = selectedDays,
-                    enabled = true,
-                    onEnabledChange = {})
-                onRemindersChange(listOf(newReminder) + reminders)
-                onAddNewReminderStateChange(false)
-
-            })
-
-        }
-    }
-}
-
-@Composable
-fun EditReminderCard(
-    selectedDays: Set<String>,
-    onDaySelected: (Set<String>) -> Unit,
-    checkedState: Boolean,
-    onCheckedChange: (Boolean) -> Unit,
-    setTimeOn: Boolean,
-    onSetTimeChange: (Boolean) -> Unit,
-    hour: Int,
-    minute: String,
-    isAm: Boolean,
-    onHourChange: (Int) -> Unit,
-    onMinuteChange: (String) -> Unit,
-    onAmChange: (Boolean) -> Unit,
-    addNewReminderState: Boolean = false
-) {
-
-    val elevation by animateDpAsState(
-        targetValue = if (addNewReminderState) 10.dp else 0.dp, label = String()
-    )
-    ElevatedCard(
-        elevation = CardDefaults.cardElevation(defaultElevation = elevation),
-        colors = CardDefaults.cardColors(containerColor = Color.White),
-        shape = RoundedCornerShape(20.dp),
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(20.dp)
-    ) {
-        Column(
-            modifier = Modifier.padding(vertical = 20.dp),
-            verticalArrangement = Arrangement.spacedBy(28.dp)
-        ) {
-            DaySelection(selectedDays = selectedDays, onDaySelected = {
-                onCheckedChange(false)
-                onDaySelected(it)
-            }, checkedState = checkedState, onCheckedChange = {
-                onCheckedChange(it)
-                val days = if (it) {
-                    setOf("M", "T", "W", "Th", "F", "Sa", "Su")
-                } else {
-                    emptySet()
-                }
-                onDaySelected(days)
-            })
-            SetTimeRow(setTimeOn = setTimeOn, onSetTimeChange = onSetTimeChange)
-            AnimatedVisibility(
-                visible = setTimeOn, enter = expandVertically(), exit = shrinkVertically()
-            ) {
-                TimePicker(
-                    hour = hour,
-                    minute = minute,
-                    isAm = isAm,
-                    onHourChange = onHourChange,
-                    onMinuteChange = onMinuteChange,
-                    onAmChange = onAmChange
+                onReminderSave(
+                    Reminder(
+                        time = "$hour:$minute ${if (isAm) "AM" else "PM"}",
+                        days = selectedDays,
+                        enabled = true
+                    )
                 )
+            })
 
-            }
         }
     }
 }
 
-
+/**
+ * Button that deletes the reminder
+ * @param onClick callback for when the button is clicked
+ * @sample DeleteButton(onClick = { println("Delete") })
+ */
 @Composable
-private fun DaySelection(
-    selectedDays: Set<String> = emptySet(),
-    onDaySelected: (Set<String>) -> Unit = {},
-    checkedState: Boolean = false,
-    onCheckedChange: (Boolean) -> Unit = {},
+private fun DeleteButton(
+    onClick: () -> Unit
 ) {
-    Column(
-        verticalArrangement = Arrangement.spacedBy(16.dp),
-    ) {
-        DayOfWeekBar(
-            selectedDays = selectedDays, isWorkoutReminder = true, onDaySelected = onDaySelected
-        )
-        InfoCheckboxRow(
-            checkedState = checkedState, onCheckedChange = onCheckedChange, message = "Every Day"
-        )
-
-    }
-}
-
-@Composable
-private fun TimePicker(
-    hour: Int = 9,
-    minute: String = "00",
-    isAm: Boolean = true,
-    onHourChange: (Int) -> Unit = {},
-    onMinuteChange: (String) -> Unit = {},
-    onAmChange: (Boolean) -> Unit = {}
-) {
-    val hours = (1..12).toMutableList()
-    val minutes = (0..55 step 5).map {
-            it.toString().padStart(2, '0')
-        }.toMutableList()
-    Row(
-        modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.spacedBy(
-            20.dp, Alignment.CenterHorizontally
-        ), verticalAlignment = Alignment.CenterVertically
+    Surface(
+        shape = RoundedCornerShape(38.dp),
+        modifier = Modifier
+            .width(120.dp)
+            .height(40.dp),
+        color = Color.White,
+        shadowElevation = 6.dp,
+        onClick = onClick
     ) {
         Row(
-            horizontalArrangement = Arrangement.spacedBy(10.dp),
+            modifier = Modifier
+                .fillMaxWidth(),
+            horizontalArrangement = Arrangement.Center,
             verticalAlignment = Alignment.CenterVertically
         ) {
-            ScrollWheel(width = 70.dp,
-                itemHeight = 40.dp,
-                items = hours,
-                initialItem = hour,
-                textStyle = TextStyle(fontSize = 24.sp),
-                textColor = GRAY08,
-                selectedTextColor = Color.Black,
-                onItemSelected = { i, _ ->
-                    onHourChange(hours[i])
-                })
-            ScrollWheel(width = 70.dp,
-                itemHeight = 40.dp,
-                items = minutes,
-                initialItem = minute,
-                textStyle = TextStyle(fontSize = 24.sp),
-                textColor = GRAY08,
-                selectedTextColor = Color.Black,
-                onItemSelected = { i, _ ->
-                    onMinuteChange(minutes[i])
-                })
-
-        }
-        AmPmSelector(
-            initialIsAm = isAm, onValueChanged = onAmChange
-        )
-
-    }
-
-}
-
-@Composable
-private fun SetTimeRow(
-    setTimeOn: Boolean, onSetTimeChange: (Boolean) -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(32.dp)
-            .padding(horizontal = 30.dp),
-        horizontalArrangement = Arrangement.spacedBy(16.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            text = "Set Time",
-            fontFamily = montserratFamily,
-            fontSize = 14.sp,
-            color = GRAY04,
-            fontWeight = FontWeight.Bold
-        )
-        ReminderSwitch(checked = setTimeOn, onCheckedChange = onSetTimeChange)
-        AnimatedVisibility(visible = !setTimeOn) {
-            Text(
-                text = "Default is 9:00 am",
-                fontFamily = montserratFamily,
-                fontSize = 12.sp,
-                color = GRAY04,
-                fontWeight = FontWeight.Light
+            Icon(
+                painter = painterResource(id = R.drawable.ic_trash),
+                contentDescription = "Delete",
             )
-
+            Text(
+                text = "Delete",
+                fontFamily = montserratFamily,
+                fontSize = 14.sp,
+                color = PRIMARY_BLACK,
+                fontWeight = FontWeight.Bold,
+                modifier = Modifier.padding(start = 5.dp)
+            )
         }
-
     }
 }
 
+/**
+ * Button that submits the reminder
+ * @param onClick callback for when the button is clicked
+ * @sample SubmitReminderButton(onClick = { println("Submit") })
+ */
 @Composable
-private fun InfoCheckboxRow(
-    checkedState: Boolean, onCheckedChange: (Boolean) -> Unit, message: String
+private fun SubmitReminderButton(
+    onClick: () -> Unit
 ) {
-    Row(
+    Surface(
+        shape = RoundedCornerShape(38.dp),
         modifier = Modifier
-            .height(24.dp)
-            .padding(start = 16.dp),
-        verticalAlignment = Alignment.CenterVertically
+            .width(120.dp)
+            .height(40.dp),
+        color = PRIMARY_BLACK,
+        shadowElevation = 6.dp,
+        onClick = onClick
     ) {
-        Checkbox(
-            checked = checkedState,
-            onCheckedChange = onCheckedChange,
-            colors = CheckboxDefaults.colors(
-                checkedColor = PRIMARY_YELLOW,
-                checkmarkColor = Color.Black,
-                uncheckedColor = GRAY04,
-            ),
-            modifier = Modifier.clip(RoundedCornerShape(4.dp))
-        )
-        Text(
-            text = message,
-            fontSize = 14.sp,
-            fontFamily = montserratFamily,
-            color = GRAY04,
-            fontWeight = FontWeight.SemiBold
-        )
+        Row(
+            modifier = Modifier
+                .fillMaxWidth(),
+            horizontalArrangement = Arrangement.Center,
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Icon(
+                painter = painterResource(id = R.drawable.ic_check),
+                contentDescription = "Done",
+                tint = Color.White
+            )
+            Text(
+                text = "Done",
+                fontFamily = montserratFamily,
+                fontSize = 14.sp,
+                color = Color.White,
+                fontWeight = FontWeight.Bold,
+                modifier = Modifier.padding(start = 5.dp)
+            )
+        }
     }
 }
+
